+++ docs/release_notes.rst
@@ -5,6 +5,8 @@ bpack v0.8.0 (UNRELEASED)
 -------------------------
 
 * Improved User Guide
+* :func:`bpack.np.unpackbits` now supports ``bits_per sample`` values
+  for 1 to 8 and 10, 12, 14, 16, 32, 64.
 
 
 bpack v0.7.0 (21/01/2021)